tfm_ingestor/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tfm-ingestor"
version = "0.1.0"
description = "TFM: Enriquecimiento de metadatos de gobierno en OpenMetadata (DCAT-like) via API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0",
  "requests>=2.31",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
]

[project.scripts]
tfm-ingestor = "tfm_ingestor.main:cli"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

